+++ packages/wonder-blocks-link/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @khanacademy/wonder-blocks-link
 
+## 7.0.4
+
+### Patch Changes
+
+-   11a0f5c6: No functional changes. Adding prepublishOnly script.
+-   Updated dependencies [11a0f5c6]
+    -   @khanacademy/wonder-blocks-clickable@5.0.4
+    -   @khanacademy/wonder-blocks-tokens@3.0.1
+    -   @khanacademy/wonder-blocks-core@11.0.1
+    -   @khanacademy/wonder-blocks-icon@5.0.4
+
 ## 7.0.3
 
 ### Patch Changes